devop: create gasPrice mixin
